fix: /api/puzzle/puzzleid/save-progress 사이즈 버그 해결, 위치 int에서 double로 변경

+++ lib/services/puzzle/dto/puzzle_save_progress/puzzlepiece_position.dart
@@ -4,8 +4,8 @@ part 'puzzlepiece_position.g.dart';
 
 @JsonSerializable()
 class PuzzlePiecePosition {
-  final int row;
-  final int col;
+  final double row;
+  final double col;
 
   PuzzlePiecePosition({
     required this.row,
